### 2024-09-28 #661 @WeichengSun
ILSM: Analyze Interconnection Structure of Multilayer Interaction Networks
Submitting Author Name: Weicheng Sun
Submitting Author Github Handle: @WeichengSun
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/WeichengSun/ILSM
Version submitted: 1.0.3.2
Submission type: Standard

Version accepted: TBD

---

```
Package: ILSM
Title: Analyze Interconnection Structure of Multilayer Interaction Networks
Version: 1.0.3.2
    person("Weicheng", "Sun", , "sunwch2023@lzu.edu.cn", role = c("aut", "cre")),
    person("Chuan", "Yan", , "yanchuan@lzu.edu.cn", role = c("aut")),
    person("Yangyang", "Zhao", , "120220907711@lzu.edu.cn", role = c("aut")))
    In view of the analysis of the structural characteristics of the multilayer
    network has been complete, however, there is still a lack of a unified
    operation that can quickly obtain the corresponding characteristics of the
    multilayer network.
    To solve this insufficiency, 'ILSM' was designed for supporting calculating
    such metrics of multilayer networks by functions of this R package.
Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    Matrix,
    igraph

```

This package that wrap professional programs used for multilayer interaction network research. These programs must be specific to interconnection structure.  It focuses on interacnnection motif for handling matrices or 'igraph'data.

In community ecology, ecological network is used to represent the complex ecological interaction system. Therefore, researchers in this field often pay attention to the work from the network perspective. Therefore, this package is based on the multilayer ecological network to study the interconnection structure, hoping to provide convenience for the mentioned audience.

There is an R package(bmotif) that does similar work in the field of ecological networks, but the program provided by it is only used to study the structure of bipartite networks and is relatively simple, while the new package only studies the structure of multilayer interacting networks, which is more complex.

n/a

[#654](https://github.com/ropensci/software-review/issues/654)

Because we have submitted the package ILSM to CRAN ahead of time, the package name is duplicated and the  [`pkgcheck` items](https://github.com/WeichengSun/ILSM/actions/runs/11073028649) result is not available.
