### 2024-09-29 #661 @WeichengSun
Hi @adamhsparks, I can't fix this check error: ✖️ Package name is not available (on CRAN).
Because the ILSM on CRAN and this submission here are all my projects
