### 2024-09-29 #661 @adamhsparks
yes, I'm aware, I checked your README and CRAN. Thanks for clarifying though!
