### 2024-09-29 #661 @adamhsparks
Hi @WeichengSun, there's a few NOTEs that I've found when I run checks locally, could you please work to address them as I find a handling editor for {ILSM}?

```r
❯ checking top-level files ... NOTE
  Non-standard files/directories found at top level:
    ‘Contributing.Rmd’ ‘README.Rmd’ ‘codemeta.json’

❯ checking Rd files ... [0s/0s] NOTE
  checkRd: (-3) coid.Rd:32: Lost braces
      32 | \itemize{\item{\code{network.or.subnet_mat1}: input a 'igraph' of network data independently or input sparse matrix together with \code{subnet_mat2}.}}
         |               ^
  checkRd: (-3) coid.Rd:35: Lost braces
      35 | \itemize{\item{\code{network.or.subnet_mat1}: must input matrix(or data.frame) together with \code{subnet_mat2}. the matrix can be sparse matrix and matrix of interaction strength.}}
         |               ^
  checkRd: (-3) coid.Rd:41: Lost braces
      41 | \item{(1). Input in a network of type "igraph" alone.}
         |      ^
  checkRd: (-3) coid.Rd:42: Lost braces
      42 | \item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
         |      ^
  checkRd: (-3) coid.Rd:47: Lost braces
      47 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) coid.Rd:49: Lost braces
      49 | \item{When the two matrices can have different numbers of rows:}
         |      ^
  checkRd: (-3) coid.Rd:51: Lost braces
      51 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) coid.Rd:52: Lost braces
      52 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) coid.Rd:55: Lost braces
      55 | \item{When the two matrices can have the same numbers of rows:}
         |      ^
  checkRd: (-3) coid.Rd:57: Lost braces
      57 | \item{No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).}
         |      ^
  checkRd: (-3) coid.Rd:60: Lost braces
      60 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^
  checkRd: (-3) cois.Rd:32: Lost braces
      32 | \itemize{\item{\code{network.or.subnet_mat1}: input a 'igraph' of network data independently or input sparse matrix together with \code{subnet_mat2}.}}
         |               ^
  checkRd: (-3) cois.Rd:35: Lost braces
      35 | \itemize{\item{\code{network.or.subnet_mat1}: must input matrix(or data.frame) together with \code{subnet_mat2}. the matrix can be sparse matrix and matrix of interaction strength.}}
         |               ^
  checkRd: (-3) cois.Rd:41: Lost braces
      41 | \item{(1). Input in a network of type "igraph" alone.}
         |      ^
  checkRd: (-3) cois.Rd:42: Lost braces
      42 | \item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
         |      ^
  checkRd: (-3) cois.Rd:47: Lost braces
      47 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) cois.Rd:49: Lost braces
      49 | \item{When the two matrices can have different numbers of rows:}
         |      ^
  checkRd: (-3) cois.Rd:51: Lost braces
      51 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) cois.Rd:52: Lost braces
      52 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) cois.Rd:55: Lost braces
      55 | \item{When the two matrices can have the same numbers of rows:}
         |      ^
  checkRd: (-3) cois.Rd:57: Lost braces
      57 | \item{No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).}
         |      ^
  checkRd: (-3) cois.Rd:60: Lost braces
      60 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^
  checkRd: (-3) hc.Rd:29: Lost braces
      29 | \item{(1). Input in a network of type "igraph" alone.}
         |      ^
  checkRd: (-3) hc.Rd:30: Lost braces
      30 | \item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
         |      ^
  checkRd: (-3) hc.Rd:35: Lost braces
      35 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) hc.Rd:36: Lost braces
      36 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) hc.Rd:37: Lost braces
      37 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) hc.Rd:38: Lost braces
      38 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^
  checkRd: (-3) icmotif_count.Rd:31: Lost braces
      31 | \item{(1). Input in a network of type "igraph" alone.}
         |      ^
  checkRd: (-3) icmotif_count.Rd:32: Lost braces
      32 | \item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
         |      ^
  checkRd: (-3) icmotif_count.Rd:37: Lost braces
      37 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) icmotif_count.Rd:39: Lost braces
      39 | \item{When the two matrices can have different numbers of rows:}
         |      ^
  checkRd: (-3) icmotif_count.Rd:41: Lost braces
      41 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) icmotif_count.Rd:42: Lost braces
      42 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) icmotif_count.Rd:45: Lost braces
      45 | \item{When the two matrices can have the same numbers of rows:}
         |      ^
  checkRd: (-3) icmotif_count.Rd:47: Lost braces
      47 | \item{No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).}
         |      ^
  checkRd: (-3) icmotif_count.Rd:50: Lost braces
      50 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^
  checkRd: (-3) icmotif_role.Rd:31: Lost braces
      31 | \item{(1). Input in a network of type "igraph" alone.}
         |      ^
  checkRd: (-3) icmotif_role.Rd:32: Lost braces
      32 | \item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
         |      ^
  checkRd: (-3) icmotif_role.Rd:37: Lost braces
      37 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) icmotif_role.Rd:39: Lost braces
      39 | \item{When the two matrices can have different numbers of rows:}
         |      ^
  checkRd: (-3) icmotif_role.Rd:41: Lost braces
      41 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) icmotif_role.Rd:42: Lost braces
      42 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) icmotif_role.Rd:45: Lost braces
      45 | \item{When the two matrices can have the same numbers of rows:}
         |      ^
  checkRd: (-3) icmotif_role.Rd:47: Lost braces
      47 | \item{No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).}
         |      ^
  checkRd: (-3) icmotif_role.Rd:50: Lost braces
      50 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^
  checkRd: (-3) igraph_from_matrices.Rd:31: Lost braces
      31 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) igraph_from_matrices.Rd:33: Lost braces
      33 | \item{When the two matrices can have different numbers of rows:}
         |      ^
  checkRd: (-3) igraph_from_matrices.Rd:35: Lost braces
      35 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) igraph_from_matrices.Rd:36: Lost braces
      36 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) igraph_from_matrices.Rd:39: Lost braces
      39 | \item{When the two matrices can have the same numbers of rows:}
         |      ^
  checkRd: (-3) igraph_from_matrices.Rd:41: Lost braces
      41 | \item{No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).}
         |      ^
  checkRd: (-3) igraph_from_matrices.Rd:44: Lost braces
      44 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^
  checkRd: (-3) node_cv.Rd:47: Lost braces
      47 | \item{(1). Input in a network of type "igraph" alone.}
         |      ^
  checkRd: (-3) node_cv.Rd:48: Lost braces
      48 | \item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
         |      ^
  checkRd: (-3) node_cv.Rd:53: Lost braces
      53 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) node_cv.Rd:55: Lost braces
      55 | \item{When the two matrices can have different numbers of rows:}
         |      ^
  checkRd: (-3) node_cv.Rd:57: Lost braces
      57 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) node_cv.Rd:58: Lost braces
      58 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) node_cv.Rd:61: Lost braces
      61 | \item{When the two matrices can have the same numbers of rows:}
         |      ^
  checkRd: (-3) node_cv.Rd:63: Lost braces
      63 | \item{No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).}
         |      ^
  checkRd: (-3) node_cv.Rd:66: Lost braces
      66 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^
  checkRd: (-3) node_cv.Rd:34-35: Lost braces
      34 | \item{If \code{type} is either of "degree", "pagerank", "hub", "authority", "katz", "eigenvector", "closeness", the data frame has two columns, and the second column corresponds to either of "Degree", "Pagerank_versatility",
         |      ^
  checkRd: (-3) node_cv.Rd:36-37: Lost braces
      36 | \item{If \code{type} is "all", the data frame has eight columns, and columns form the second to the eighth correspond to "Degree", "Pagerank_versatility",
         |      ^
  checkRd: (-3) null_model.Rd:30: Lost braces
      30 | \item{For each of the four types of null models, there are corresponding algorithms. The first type, “subnetwork1”, involved scrambling the adjacency matrix of the first and second groups of the multilayer network.}
         |      ^
  checkRd: (-3) null_model.Rd:31: Lost braces
      31 | \item{The second type, “subnetwork2”, focused on scrambling the adjacency matrix of the second and third groups. }
         |      ^
  checkRd: (-3) null_model.Rd:32: Lost braces
      32 | \item{Comprehensively, the third type, “all”, blended the approaches of the first two to disarrange the entire network's adjacency matrix, achieving a thorough perturbation of the network's structure. }
         |      ^
  checkRd: (-3) null_model.Rd:33: Lost braces
      33 | \item{The last type named “Savue” that disarranged inherent structure in terms of the groups of species connected by each interconnecting species of every subnetworks, thus exhibiting different interconnection patterns.}}
         |      ^
  checkRd: (-3) pc.Rd:29: Lost braces
      29 | \item{(1). Input in a network of type "igraph" alone.}
         |      ^
  checkRd: (-3) pc.Rd:30: Lost braces
      30 | \item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
         |      ^
  checkRd: (-3) pc.Rd:35: Lost braces
      35 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) pc.Rd:36: Lost braces
      36 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) pc.Rd:37: Lost braces
      37 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) pc.Rd:38: Lost braces
      38 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^
  checkRd: (-3) poc.Rd:29: Lost braces
      29 | \item{(1). Input in a network of type "igraph" alone.}
         |      ^
  checkRd: (-3) poc.Rd:30: Lost braces
      30 | \item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
         |      ^
  checkRd: (-3) poc.Rd:35: Lost braces
      35 | \item{Try to make the rows of both matrices have the same attributes. Or we default:}
         |      ^
  checkRd: (-3) poc.Rd:36: Lost braces
      36 | \item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) poc.Rd:37: Lost braces
      37 | \item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
         |      ^
  checkRd: (-3) poc.Rd:38: Lost braces
      38 | \item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
         |      ^

0 errors ✔ | 0 warnings ✔ | 2 notes ✖
```
