### 2024-09-29 071a306
Update coid.R
