### 2024-10-09 fd3e961
Update coid.Rd
