### 2024-10-09 96ffc75
Update hc.Rd
