### 2024-10-09 3af70e7
Update icmotif_count.Rd
