+++ man/icmotif_role.Rd
@@ -24,30 +24,30 @@ Each row of matrix corresponds to a interconnecting species in the second layer 
 Counting the number of 70 roles about interconnecting species in multilayer network motifs.
 }
 \details{
-\strong{network.or.subnet_mat1} and \strong{subnet_mat2}
+\strong{network.or.subnet_mat1 and subnet_mat2}
 
 There are two types of \code{network.or.subnet_mat1} that can be processed:
 \itemize{
-\item{(1). Input in a network of type "igraph" alone.}
-\item{(2). Must be entered as data frame or matrix with \code{subnet_mat2}.}
+\item Input in a network of type "igraph" alone.
+\item Must be entered as data frame or matrix with \code{subnet_mat2}.
 }
 
 If the type of inputting is data frame or matrix, please make sure the row of \code{network.or.subnet_mat1} and \code{subnet_mat2} correspond with the second group of species that both belong to two subnetworks and interact with other groups of species.
 \itemize{
-\item{Try to make the rows of both matrices have the same attributes. Or we default:}
+\item Try to make the rows of both matrices have the same attributes. Or we default:
 
-\item{When the two matrices can have different numbers of rows:}
+\item When the two matrices can have different numbers of rows:
 \itemize{
-\item{(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.}
-\item{(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.}
+\item 1. If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.
+\item 2. If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.
 }
 
-\item{When the two matrices can have the same numbers of rows:}
+\item When the two matrices can have the same numbers of rows:
 \itemize{
-\item{No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).}
+\item No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).
 }
 
-\item{The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!}
+\item The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!
 }
 About a network of type "igraph", It can be obtained from the connection matrices of subnetworks by the function \code{igraph_from_matrices}.
 }
