### 2024-10-09 4ed4204
Update igraph_from_matrices.Rd
