+++ man/null_model.Rd
@@ -27,10 +27,11 @@ The null model could be generated according to different matrix scrambling algor
 \strong{null_type}
 
 \itemize{
-\item{For each of the four types of null models, there are corresponding algorithms. The first type, “subnetwork1”, involved scrambling the adjacency matrix of the first and second groups of the multilayer network.}
-\item{The second type, “subnetwork2”, focused on scrambling the adjacency matrix of the second and third groups. }
-\item{Comprehensively, the third type, “all”, blended the approaches of the first two to disarrange the entire network's adjacency matrix, achieving a thorough perturbation of the network's structure. }
-\item{The last type named “Savue” that disarranged inherent structure in terms of the groups of species connected by each interconnecting species of every subnetworks, thus exhibiting different interconnection patterns.}}
+\item For each of the four types of null models, there are corresponding algorithms. The first type, “subnetwork1”, involved scrambling the adjacency matrix of the first and second groups of the multilayer network.
+\item The second type, “subnetwork2”, focused on scrambling the adjacency matrix of the second and third groups.
+\item Comprehensively, the third type, “all”, blended the approaches of the first two to disarrange the entire network's adjacency matrix, achieving a thorough perturbation of the network's structure.
+\item The last type named “Savue” that disarranged inherent structure in terms of the groups of species connected by each interconnecting species of every subnetworks, thus exhibiting different interconnection patterns.
+}
 
 \strong{network}
 
