### 2024-10-09 280a5f7
Update hc.R
