+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   motif_guideline: motif_guideline.html
-last_built: 2024-09-27T15:05Z
+last_built: 2024-10-09T03:08Z
