### 2024-10-09 #661 @adamhsparks
Thanks, @WeichengSun,

1. I still see an issue with missing braces as shown here:

```r
  ✖ checking Rd files ... [0s/0s] NOTE checkRd: (-3)
    node_cv.Rd:34-35: Lost braces 34 | \item{If \code{type} is either
    of "degree", "pagerank", "hub", "authority", "katz", "eigenvector",
    "closeness", the data frame has two columns, and the second column
    corresponds to either of "Degree", "Pagerank_versatility", | ^
    checkRd: (-3) node_cv.Rd:36-37: Lost braces 36 | \item{If
    \code{type} is "all", the data frame has eight columns, and columns
    form the second to the eighth correspond to "Degree",
    "Pagerank_versatility", | ^
```

2. There is one other issue I'd like to have addressed. This doesn't appear to be a stats package, but you have SRR roclets. Could you please remove those as well as they are unnecessary and have already caused some confusion (mainly me being confused about this being a stats or normal package submission).

Thanks!
