### 2024-10-09 #661 @WeichengSun
OK @adamhsparks, I will continue to complete the issue, and I also accept your advice to put it on hold. Since this is only a tool for our scientific research, but our research work still needs to continue. Please inform me and provide other relevant requirements after you finish the working on a category for network standards. Thank you.
Mutual encouragement.
