+++ NAMESPACE
@@ -9,6 +9,7 @@ export(icmotif_count)
 export(icmotif_role)
 export(igraph_from_matrices)
 export(node_cv)
+export(node_icc)
 export(null_model)
 export(pc)
 export(poc)
@@ -21,10 +22,17 @@ importFrom(Matrix,sparseMatrix)
 importFrom(igraph,V)
 importFrom(igraph,as.directed)
 importFrom(igraph,as_adjacency_matrix)
+importFrom(igraph,authority_score)
+importFrom(igraph,betweenness)
+importFrom(igraph,centr_eigen)
+importFrom(igraph,closeness)
+importFrom(igraph,degree)
 importFrom(igraph,distances)
 importFrom(igraph,get.edgelist)
 importFrom(igraph,graph_from_adjacency_matrix)
 importFrom(igraph,graph_from_edgelist)
+importFrom(igraph,hub_score)
 importFrom(igraph,layout_with_sugiyama)
+importFrom(igraph,page_rank)
 importFrom(igraph,plot.igraph)
 importFrom(stats,runif)
