### 2024-10-31 22d3860
Create node_icc.R
