+++ R/node_icc.R
@@ -1,0 +1,209 @@
+#' Measuring Interconnection Centrality of Tripartite Interaction Network
+#'
+#' The Interconnection Centrality of nodes is revealed by several centrality
+#' measures that have now been applied to tripartite networks, such as Degree,
+#' Pagerank, Hub, Authority, Eigenvector, and Closeness Betweenness centrality.
+#'
+#' @param network.or.subnet_mat1 Either a tripartite (multilayer) network of
+#' 'igraph' class which contains three groups of species and interactions within
+#'  layers without interactions between each group of species, or a numeric
+#'  matrix(or data.frame) representing interactions between two groups of
+#'  species.
+#'  Each row and column of matrix represents single species in the second and
+#'  first groups of the tripartite network respectively.
+#'  Elements of matrix are non-zero numbers if the two groups of species are
+#'  connected, and 0 otherwise.
+#'
+#' @param subnet_mat2 A numeric matrix(or data.frame) representing interactions
+#' between two groups of species.
+#'  Each row and column of matrix represents single species in the second and
+#'  third groups of the tripartite network respectively.
+#'  Elements of matrix are non-zero numbers if the two groups of species are
+#'  connected, and 0 otherwise. If \code{network.or.subnet_mat1} is "igraph",
+#'  \code{subnet_mat2} defaults to NULL.
+#'
+#' @param isDirected1 Logical. Whether the interaction between the two groups of
+#'  species in \code{mat1} is unidirectional.Default to TRUE, such as Predation
+#'  and Herbivory. Otherwise it is bidirectional, such as Mutualism.
+#' @param isDirected2 Logical. Whether the interaction between the two groups of
+#'  species in \code{mat2} is unidirectional.Default to TRUE, such as Predation
+#'  and Herbivory. Otherwise it is bidirectional, such as Mutualism.
+#' @param type Character. Including "degree", "pagerank", "hub", "authority",
+#' "eigenvector", "closeness", "betweenness", and "all".
+#'
+#' @details
+#'
+#' \strong{network.or.subnet_mat1 and subnet_mat2}
+#'
+#' There are two types of \code{network.or.subnet_mat1} that can be processed:
+#' \itemize{
+#' \item Input in a network of type "igraph" alone.
+#' \item Must be entered as data frame or matrix with \code{subnet_mat2}.
+#' }
+#'
+#' If the type of inputting is data frame or matrix, please make sure the row of
+#'  \code{network.or.subnet_mat1} and \code{subnet_mat2} correspond with the
+#'  second group of species that both belong to two subnetworks and interact
+#'  with other groups of species.
+#' \itemize{
+#' \item Try to make the rows of both matrices have the same attributes. Or we
+#' default:
+#'
+#' \item When the two matrices can have different numbers of rows:
+#' \itemize{
+#' \item 1. If both matrices have row names, then the function counts all row
+#' names to produce two new matrices with the same row names.
+#' \item 2. If at most one matrix has row names, the function assigns new row
+#' names to both matrices on a row-to-row basis (any extra row names are
+#' assigned a new value) and then counts all row names to produce two new
+#' matrices with the same row names.
+#' }
+#'
+#' \item When the two matrices can have the same numbers of rows:
+#' \itemize{
+#' \item No matter how the row names of the two matrices are arranged, as long
+#' as the row names are exactly the same; But we don't handle matrices with
+#' empty row names (the function will give an error).
[…]
