### 2024-10-31 e86b77c
Create test-node_icc.R
