+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   motif_guideline: motif_guideline.html
-last_built: 2024-10-09T13:52Z
+last_built: 2024-10-31T09:35Z
