### 2024-11-01 b2d9bb5
Update node_cv.Rd
