+++ README.Rmd
@@ -15,16 +15,16 @@ knitr::opts_chunk$set(
 
 <!-- badges: start -->
 
-[![R-CMD-check](https://github.com/WeichengSun/ILSM/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/WeichengSun/ILSM/actions/workflows/R-CMD-check.yaml) [![Project Status: Inactive -- The project has reached a stable, usable state but is no longer being actively developed; support/maintenance will be provided as time allows.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/ILSM)](https://www.r-pkg.org/pkg/ILSM) [![Codecov test coverage](https://codecov.io/gh/WeichengSun/ILSM/graph/badge.svg)](https://app.codecov.io/gh/WeichengSun/ILSM) [![Lifecycle: stable](https://img.shields.io/badge/lifecycle-Stable-blue.svg)](https://lifecycle.r-lib.org/articles/stages.html##Stable)
+[![R-CMD-check](https://github.com/WeichengSun/ILSM/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/WeichengSun/ILSM/actions/workflows/R-CMD-check.yaml) [![Project Status: Active -- The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![CRAN_Status](http://www.r-pkg.org/badges/version/ILSM)](https://cran.r-project.org/package=ILSM) [![Codecov test coverage](https://codecov.io/gh/WeichengSun/ILSM/graph/badge.svg)](https://app.codecov.io/gh/WeichengSun/ILSM) [![Lifecycle: stable](https://img.shields.io/badge/lifecycle-Stable-blue.svg)](https://lifecycle.r-lib.org/articles/stages.html##Stable) [![R-CMD-check](https://github.com/WeichengSun/ILSM/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/WeichengSun/ILSM/actions/workflows/R-CMD-check.yaml)
 
 <!-- badges: end -->
 
 ILSM is designed for analyzing interconnection structures including interconnection patterns, versatility and motifs in multilayer interaction networks.
 
-The figure below shows an example multilayer interaction network (a, with three groups of species and two interaction layers, showing two common forms) and interconnection structures (b, interconnection pattern, c, interconnection motif, d, versatility). Different colors of nodes indicate different groups of species, and the dashed lines indicate the connector species. In panel a, the connector species in the left panel were indicated by dashed links, and those in the right panel have links from both layers.
+The figure below shows an example multilayer interaction network (a, with three groups of species and two interaction layers, showing two common forms) and interconnection structures (b, interconnection pattern, c, interconnection motif, d, interconnection centratility). Different colors of nodes indicate different groups of species, and the dashed lines indicate the connector species. In panel a, the connector species in the left panel were indicated by dashed links, and those in the right panel have links from both layers.
 
 ```{r echo=F, out.width = "100%"}
-knitr::include_graphics("./man/figure/combine.png")
+knitr::include_graphics("./man/figure/pattern.png")
 ```
 
 ## Installation
@@ -70,8 +70,8 @@ cois(Net)
 pc(Net)
 hc(Net)
 
-## connector nodes' versatility
-node_cv(Net)
+## connector nodes' interconnection centratility
+node_icc(Net)
 ```
 
 ## License
