+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ILSM
 Title: Analyze Interconnection Structure of Tripartite Interaction Networks
-Version: 1.0.3.2
+Version: 1.0.3.3
 Authors@R: c(
     person("Weicheng", "Sun", , "sunwch2023@lzu.edu.cn", role = c("aut", "cre")),
     person("Chuan", "Yan", , "yanchuan@lzu.edu.cn", role = c("aut")),
