### 2024-11-05 3abee7b
Update motif_guideline.R
