+++ inst/doc/motif_guideline.R
@@ -5,7 +5,7 @@ knitr::opts_chunk$set(
 )
 
 ## ----echo=F, out.width = "100%"-----------------------------------------------
-knitr::include_graphics("../man/figure/combine.png")
+knitr::include_graphics("../man/figure/pattern.png")
 
 ## ----echo = T,eval = F--------------------------------------------------------
 #  library(ILSM)
