### 2024-11-05 5da1c67
Update motif_guideline.Rmd
