+++ man/node_cv.Rd
@@ -1,113 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/node_cv.R
-\name{node_cv}
-\alias{node_cv}
-\title{Measuring node versatility of multilayer network}
-\usage{
-node_cv(
-  network.or.subnet_mat1,
-  subnet_mat2 = NULL,
-  isDirected1 = TRUE,
-  isDirected2 = TRUE,
-  type = c("degree", "pagerank", "hub", "authority", "katz", "eigenvector", "closeness",
-    "all")
-)
-}
-\arguments{
-\item{network.or.subnet_mat1}{Either a multilayer(tripartite) network of 'igraph' class which contains three groups of species and interactions within layers without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
-Each row and column of matrix represents single species in the second and first groups of the tripartite network respectively.
-Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise.}
-
-\item{subnet_mat2}{A numeric matrix(or data.frame) representing interactions between two groups of species.
-Each row and column of matrix represents single species in the second and third groups of the tripartite network respectively.
-Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise. If \code{network.or.subnet_mat1} is "igraph", \code{subnet_mat2} defaults to NULL.}
-
-\item{isDirected1}{Logical. Whether the interaction between the two groups of species in \code{mat1} is unidirectional.Default to TRUE, such as Predation and Herbivory. Otherwise it is bidirectional, such as Mutualism.}
-
-\item{isDirected2}{Logical. Whether the interaction between the two groups of species in \code{mat2} is unidirectional.Default to TRUE, such as Predation and Herbivory. Otherwise it is bidirectional, such as Mutualism.}
-
-\item{type}{Character. Including "degree", "pagerank", "hub", "authority", "katz", "eigenvector", "closeness", and "all".}
-}
-\value{
-Return a data frame with the first row "node" for each node of network representing each species.
-\itemize{
-\item If \code{type} is either of "degree", "pagerank", "hub", "authority", "katz", "eigenvector", "closeness", the data frame has two columns, and the second column corresponds to either of "Degree", "Pagerank_versatility",
-"Hub_versatility", "Authority_versatility", "Katz_versatility", "Eigenvector_versatility", "Closeness_versatility".
-\item If \code{type} is "all", the data frame has eight columns, and columns form the second to the eighth correspond to "Degree", "Pagerank_versatility",
-"Hub_versatility", "Authority_versatility", "Katz_versatility", "Eigenvector_versatility", "Closeness_versatility".
-}
-}
-\description{
-The \code{node_icc} will replace the function \code{node_cv} in tripartite
-network.
-We will temporarily suspend the development of node versatility within the
-multilayer network, but we will continue to move forward with the project.
-}
-\details{
-The versatility of nodes is revealed by several centrality measures that have now been applied to multilayer networks, such as Degree, Pagerank, Hub, Authority, Katz, Eigenvector, and Closeness centrality.
-
-\strong{network.or.subnet_mat1 and subnet_mat2}
-
-There are two types of \code{network.or.subnet_mat1} that can be processed:
-\itemize{
-\item Input in a network of type "igraph" alone.
-\item Must be entered as data frame or matrix with \code{subnet_mat2}.
-}
-
[…]
