+++ NAMESPACE
@@ -8,7 +8,6 @@ export(hc)
 export(icmotif_count)
 export(icmotif_role)
 export(igraph_from_matrices)
-export(node_cv)
 export(node_icc)
 export(null_model)
 export(pc)
