### 2024-11-05 6d81633
Update NEWS.md
