### 2024-11-05 a710ce2
Delete node_cv.R
