+++ README.Rmd
@@ -19,7 +19,7 @@ knitr::opts_chunk$set(
 
 <!-- badges: end -->
 
-ILSM is designed for analyzing interconnection structures including interconnection patterns, versatility and motifs in multilayer interaction networks.
+ILSM is designed for analyzing interconnection structures including interconnection patterns, centrality and motifs in multilayer interaction networks.
 
 The figure below shows an example multilayer interaction network (a, with three groups of species and two interaction layers, showing two common forms) and interconnection structures (b, interconnection pattern, c, interconnection motif, d, interconnection centratility). Different colors of nodes indicate different groups of species, and the dashed lines indicate the connector species. In panel a, the connector species in the left panel were indicated by dashed links, and those in the right panel have links from both layers.
 
