### 2024-11-07 653a87f
Update node_icc.R
