### 2024-11-07 b26f19f
Update test-cois.R
