### 2024-11-08 8db95a3
Update motif_guideline.Rmd
