+++ inst/doc/motif_guideline.Rmd
@@ -14,13 +14,13 @@ knitr::opts_chunk$set(
 )
 ```
 
-The visualization of an example multilayer interaction network with three parties of nodes and two interaction layers (a, two common forms in literature) and interconnection structures (b, interconnection pattern, c, interconnection motif, d, interconnection centrality). Different colors of nodes indicate different parties of species, and the dashed lines indicate the shared species. (a) The shared species in the left panel were indicated by dashed links, and those in the right panel have links from both layers.
+The visualization of an example tripartite interaction network with three parties of nodes and two interaction layers (a, two common forms in literature) and interconnection structures (b, interconnection pattern, c, interconnection motif, d, interconnection centrality). Different colors of nodes indicate different parties of species, and the dashed lines indicate the shared species. (a) The shared species in the left panel were indicated by dashed links, and those in the right panel have links from both layers.
 
 ```{r echo=F, out.width = "100%"}
 knitr::include_graphics("../man/figure/pattern.png")
 ```
 
-In multilayer network, 48 types of motifs of nodes ranging from 3 to 6 are obtained by that
+In tripartite network, 48 types of motifs of nodes ranging from 3 to 6 are obtained by that
 
 ```{r echo = T,eval = F}
 library(ILSM)
