### 2024-11-08 f9987e3
Update coid.Rd
