+++ man/coid.Rd
@@ -7,7 +7,7 @@
 coid(network.or.subnet_mat1, subnet_mat2 = NULL, weighted = FALSE)
 }
 \arguments{
-\item{network.or.subnet_mat1}{Either a multilayer(tripartite) network of 'igraph' class which contains three groups of species and interactions within layers without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
+\item{network.or.subnet_mat1}{Either a tripartite network of 'igraph' class which contains three groups of species and interactions within layers without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
 Each row and column of matrix represents single species in the second and first groups of the tripartite network respectively.
 Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise.}
 
@@ -65,7 +65,7 @@ About a network of type "igraph", It can be obtained from the connection matrice
 }
 \examples{
 
-## generate a random multilayer network
+## generate a random tripartite network
 set.seed(12)
 Net <- build_net(11,15,16,0.2)
 
