### 2024-11-08 d51a933
Update cois.Rd
