### 2024-11-08 5604ea0
Update hc.Rd
