### 2024-11-08 6a8b2cb
Update icmotif_count.Rd
