### 2024-11-08 97d21ce
Update igraph_from_matrices.Rd
