### 2024-11-08 461d00f
Update role_sim.Rd
