### 2024-11-08 5c54630
Update icmotif_count.R
