### 2024-11-08 5625aeb
Update igraph_from_matrices.R
