### 2024-11-08 8f098f0
Update node_icc.R
