+++ R/null_model.R
@@ -1,8 +1,8 @@
-#' Null model of multilayer network
+#' Null model of tripartite network
 #'
-#' The null model could be generated according to different matrix scrambling algorithms for interconnection patterns in the multilayer network
+#' The null model could be generated according to different matrix scrambling algorithms for interconnection patterns in the tripartite network
 #'
-#' @param network A multilayer(tripartite) network of 'igraph' class. The network contains three groups of species and interactions within layers without interactions between each group of species.
+#' @param network A tripartite network of 'igraph' class. The network contains three groups of species and interactions within layers without interactions between each group of species.
 #' @param number A numeric value. The number of null model.  Default to NULL representing number 1.
 #' @param null_type Logical. Four matrix scrambling algorithms. If null_type = NULL, default to "all".
 #'
@@ -11,7 +11,7 @@
 #' \strong{null_type}
 #'
 #' \itemize{
-#' \item For each of the four types of null models, there are corresponding algorithms. The first type, “subnetwork1”, involved scrambling the adjacency matrix of the first and second groups of the multilayer network.
+#' \item For each of the four types of null models, there are corresponding algorithms. The first type, “subnetwork1”, involved scrambling the adjacency matrix of the first and second groups of the tripartite network.
 #' \item The second type, “subnetwork2”, focused on scrambling the adjacency matrix of the second and third groups.
 #' \item Comprehensively, the third type, “all”, blended the approaches of the first two to disarrange the entire network's adjacency matrix, achieving a thorough perturbation of the network's structure.
 #' \item The last type named “Savue” that disarranged inherent structure in terms of the groups of species connected by each interconnecting species of every subnetworks, thus exhibiting different interconnection patterns.
@@ -23,7 +23,7 @@
 #'
 #' @return
 #'
-#' Return a list contains one or more elements. Each element represent a null model of multilayer network.
+#' Return a list contains one or more elements. Each element represent a null model of tripartite network.
 #'
 #' @references
 #' Vázquez, D. P., C. J. Melian, N. M. Williams, N. Blüthgen, B. R. Krasnov, and R. Poulin. 2007. Species abundance and asymmetric interaction strength in ecological networks. Oikos 116: 1120-1127.
@@ -39,7 +39,7 @@
 #'
 #' @examples
 #'
-#' ## generate a random multilayer network
+#' ## generate a random tripartite network
 #' set.seed(12)
 #' Net <- build_net(11,15,16,0.2)
 #'
