### 2024-11-08 e01e32d
Update role_sim.R
