+++ README.Rmd
@@ -25,9 +25,7 @@ ILSM is designed for analyzing interconnection structures including interconnect
 
 The figure below shows an example tripartite interaction network (a, with three groups of species and two interaction layers, showing two common forms) and interconnection structures (b, interconnection pattern, c, interconnection motif, d, interconnection centrality). Different colors of nodes indicate different groups of species, and the dashed lines indicate the connector species. In panel a, the connector species in the left panel were indicated by dashed links, and those in the right panel have links from both layers.
 
-```{r echo=F, out.width = "100%"}
-knitr::include_graphics("./man/figure/pattern.png")
-```
+![](man/figure/pattern.png)
 
 ## Installation
 
@@ -39,9 +37,7 @@ devtools::install_github("WeichengSun/ILSM")
 
 ## Interconnection motifs
 
-```{r echo=F, height=20}
-knitr::include_graphics("./man/figure/motif_ILSM.png")
-```
+![](man/figure/motif_ILSM.png)
 
 The figure below shows the 48 forms of interconnection motifs with 3-6 nodes. Blue and grey nodes form one layer, and grey and orange nodes form the other layer. Grey nodes are connector nodes. The motifs are named "MABC-i": M means "motif',"A" is the number of a-nodes, "B" is the number of b-nodes, "C" is the number of c-nodes and "i" is the serial number for the motifs with the same "ABC". The interconnection motifs are ordered by the number of connector nodes (from 1 to 4). The numbers from 1 to 70 in connector nodes represent the unique roles.
 
@@ -97,4 +93,3 @@ Domenico, M. D. 2022. Multilayer Networks: Analysis and Visualization. Introduct
 ## Citation
 
 Manuscript is being prepared for submission and citations are currently available at CRAN [10.32614/CRAN.package.ILSM](https://cran.r-project.org/web/packages/ILSM/index.html)
-
