### 2024-11-18 98ad10e
Update coid.Rd
