### 2024-11-18 c96eaa4
Update cois.Rd
