### 2024-11-18 4f967c0
Update icmotif_count.Rd
