### 2024-11-18 1f9ae0b
Update motif_guideline.Rmd
