### 2024-11-18 9b8e8fc
Update role_sim.R
