+++ R/pc.R
@@ -2,7 +2,7 @@
 #'
 #' Counting participation ratio that the difference in the degree of interconnecting nodes within two subnetworks.
 #'
-#' @param network.or.subnet_mat1 Either a tripartite network of 'igraph' class which contains three groups of species and interactions within layers without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
+#' @param network.or.subnet_mat1 Either a tripartite network of 'igraph' class which contains three groups of species and interactions within subnetwork without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
 #'  Each row and column of matrix represents single species in the second and first groups of the tripartite network respectively.
 #'  Elements of matrix are non-zero numbers if two groups of species are connected, and 0 otherwise.
 #' @param subnet_mat2 A numeric matrix(or data.frame) representing interactions between two groups of species.
