### 2024-11-18 9cb1fd0
Update node_icc.R
