+++ R/icmotif_role.R
@@ -2,7 +2,7 @@
 #'
 #' Counting the number of 70 roles about interconnecting species in tripartite network motifs.
 #'
-#' @param network.or.subnet_mat1 Either a tripartite network of 'igraph' class which contains three groups of species and interactions within layers without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
+#' @param network.or.subnet_mat1 Either a tripartite network of 'igraph' class which contains three groups of species and interactions within subnetwork without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
 #'  Each row and column of matrix represents single species in the second and first groups of the tripartite network respectively.
 #'  Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise.
 #'
@@ -45,7 +45,7 @@
 #' @return
 #' Returns a matrix of 70 columns representing the roles of interconnecting species in the motifs. Columns names are Role1, Role2, Role3 ... Role70.
 #'
-#' Each row of matrix corresponds to a interconnecting species in the second layer of network. If a interconnecting species is linked to both the second and third level species, the elements in this row are not all zero, otherwise the elements are all zero.
+#' Each row of matrix corresponds to a interconnecting species in the second group of network. If a interconnecting species is linked to both the first and third group species, the elements in this row are not all zero, otherwise the elements are all zero.
 #'
 #' @srrstats {G1.1} The algorithm is the first implementation of a novel algorithm.
 #'
@@ -83,7 +83,7 @@
 
 icmotif_role<-function(network.or.subnet_mat1, subnet_mat2=NULL){
    if(inherits(network.or.subnet_mat1,"igraph")==T){
-      network<-adject_net(network.or.subnet_mat1)
+      network<-adjust_net(network.or.subnet_mat1)
       PHP<-as.matrix(network[])
       PHP[PHP>0]<-1
       PH<-PHP[(V(network)$level)==0,(V(network)$level)==1]
