### 2024-11-18 6d55e15
Update icmotif_count.R
