### 2024-11-18 d9135ae
Update hc.R
