### 2024-11-18 9105ca3
Update role_sim.Rd
