+++ man/role_sim.Rd
@@ -7,7 +7,7 @@
 role_sim(spe_role_mat)
 }
 \arguments{
-\item{spe_role_mat}{A matrix of 65 columns representing the roles of
+\item{spe_role_mat}{A matrix of 70 columns representing the roles of
 interconnecting species in the motifs. Columns names are Role1, Role2,
 Role3 ... Role70.}
 }
