+++ man/poc.Rd
@@ -7,7 +7,7 @@
 poc(network.or.subnet_mat1, subnet_mat2 = NULL)
 }
 \arguments{
-\item{network.or.subnet_mat1}{Either a tripartite network of 'igraph' class which contains three groups of species and interactions within layers without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
+\item{network.or.subnet_mat1}{Either a tripartite network of 'igraph' class which contains three groups of species and interactions within subnetwork without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species.
 Each row and column of matrix represents single species in the second and first groups of the tripartite network respectively.
 Elements of matrix are non-zero numbers if two groups of species are connected, and 0 otherwise.}
 
@@ -16,10 +16,10 @@ Each row and column of matrix represents single species in the second and third 
 Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise. If \code{network.or.subnet_mat1} is "igraph", \code{subnet_mat2} defaults to NULL.}
 }
 \value{
-Print a "POC= ;" and Return a numeric value representing the proportion of sharing species in intermediate layer.
+Print a "POC= ;" and Return a numeric value representing the proportion of sharing species in connectors.
 }
 \description{
-Calculating the proportion of species sharing with other species of two subnetworks in intermediate layer.
+Calculating the proportion of species sharing with other species of two subnetworks in connectors.
 }
 \details{
 \strong{network.or.subnet_mat1 and subnet_mat2}
