+++ DESCRIPTION
@@ -17,7 +17,7 @@ URL: https://github.com/WeichengSun/ILSM
 BugReports: https://github.com/WeichengSun/ILSM/issues
 Encoding: UTF-8
 Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
-RoxygenNote: 7.3.1
+RoxygenNote: 7.3.2
 Depends: R (>= 3.5.0)
 Imports: 
     stats,
