### 2024-11-24 bbf0c68
Update icmotif_count.Rd
