### 2024-11-24 c956494
Update test-igraph_from_matrices.R
