### 2024-11-24 43a60ad
Update icmotif_count.R
