+++ NAMESPACE
@@ -14,10 +14,6 @@ export(pc)
 export(poc)
 export(role_sim)
 import(igraph)
-importFrom(Matrix,Diagonal)
-importFrom(Matrix,Matrix)
-importFrom(Matrix,drop0)
-importFrom(Matrix,sparseMatrix)
 importFrom(igraph,V)
 importFrom(igraph,as.directed)
 importFrom(igraph,as_adjacency_matrix)
@@ -26,7 +22,6 @@ importFrom(igraph,betweenness)
 importFrom(igraph,centr_eigen)
 importFrom(igraph,closeness)
 importFrom(igraph,degree)
-importFrom(igraph,distances)
 importFrom(igraph,get.edgelist)
 importFrom(igraph,graph_from_adjacency_matrix)
 importFrom(igraph,graph_from_edgelist)
