+++ R/copy_mux.R
@@ -1,242 +1,0 @@
-#' A part of code of package 'muxViz'
-#'
-#'
-#' @importFrom Matrix sparseMatrix
-#' @importFrom Matrix drop0
-#' @importFrom Matrix Diagonal
-#' @importFrom Matrix Matrix
-#' @importFrom igraph graph_from_adjacency_matrix
-#' @importFrom igraph distances
-#' @noRd
-#'
-#' @references
-#'
-#' Manlio De Domenico, Mason A. Porter, Alex Arenas, Multilayer Analysis and Visualization of Networks, published in Journal of Complex Networks 3, 159-176 (2015) (Open Access)
-#'
-#' Manlio De Domenico, Multilayer Networks: Analysis and Visualization. Introduction to muxViz with R. To be published by Springer-Verlag (2021)
-#'
-#'
-
-sumR_r<-function (A, n)
-{
-   if (n == 1) {
-      return(rbind(Matrix::colSums(A)))
-   }
-   else if (n == 2) {
-      return(cbind(Matrix::rowSums(A)))
-   }
-   else {
-      stop("ERROR! Not a valid dimension.")
-   }
-}
-
-diagR_r<- function(x, n, offset = 0) {
-   # cannot call this function just "diag", because it's an R builtin command
-   # (to-do) add check on the length of x for offset != 0
-   x <- as.numeric(x)
-   M <- matrix(0, n, n)
-   M[which(row(M) + offset == col(M))] <- x
-   return(M)
-}
-
-
-GetLargestEigenv_r<-function (Matrix)
-{
-   tmp <- eigen(Matrix, only.values = F)
-   lm <- which.max(Re(tmp$values[abs(Im(tmp$values)) < 1e-06]))
-   tmp$vectors <- tmp$vectors[, lm]
-   tmp$values <- tmp$values[lm]
-   if (all(Im(tmp$vectors) == 0)) {
-      tmp$vectors <- Re(tmp$vectors)
-   }
-   if (all(Im(tmp$values) == 0)) {
-      tmp$values <- Re(tmp$values)
-   }
-   tmp$vectors[Matrix::which(tmp$vectors > -1e-12 & tmp$vectors <
-                                1e-12)] <- 0
-   if (all(tmp$vectors[Matrix::which(tmp$vectors != 0)] < 0)) {
-      tmp$vectors <- -tmp$vectors
-   }
-   return(list(QMatrix = tmp$vectors, LMatrix = tmp$values))
-}
-
-BuildSupraAdjacencyMatrixFromExtendedEdgelist_r<-function (mEdges, Layers, Nodes, isDirected)
-{
-   if (max(max(mEdges[, 2]), max(mEdges[, 4])) != Layers) {
-      stop("Error: expected number of layers does not match the data. Aborting process.")
-   }
-   edges <- data.frame(from = mEdges[, 1] + Nodes * (mEdges[,
-                       2] - 1), to = mEdges[, 3] + Nodes * (mEdges[, 4] - 1),
-                       weight = mEdges[, 5])
-   M <- Matrix::sparseMatrix(i = edges$from, j = edges$to,
-                             x = edges$weight, dims = c(Nodes * Layers, Nodes * Layers))
-   if (sum(abs(M - Matrix::t(M))) > 1e-12 && isDirected == FALSE)
-      M <- (M + Matrix::t(M))/2
-   return(M)
-}
-
-GetMultiPageRankCentrality_r<-function (SupraAdjacencyMatrix, Layers, Nodes)
-{
-   return(GetMultiRWCentrality_r(SupraAdjacencyMatrix, Layers,
-                                 Nodes, Type = "pagerank"))
-}
-
-
-GetMultiRWCentrality_r<-function (SupraAdjacencyMatrix, Layers, Nodes, Type = "pagerank",
[…]
