### 2025-02-02 #661 @maurolepore
@WeichengSun FYI my EiC rotation just started. I see we're holding until https://github.com/ropensci/statistical-software-review-book/pull/67 is merged. I just asked for it's status at https://github.com/ropensci/statistical-software-review-book/pull/67#issuecomment-2629422637. Looking forward to continuing soon.
