### 2025-05-04 #661 @maurolepore
@WeichengSun I'm wrapping up my EiC rotation today and before I had it over to the next EiC I would like to ask if you could please write a few words explaining your intent and availability to resume.

This might weight in the priority assigned to merging this issue: https://github.com/ropensci/statistical-software-review-book/pull/67

cc' @noamross andn @mpadge
