### 2025-10-31 #661 @mpadge
Sorry @ldecicco-USGS and especially @WeichengSun, the ball is in my court on this one. I'll use your prompt to really endeavour to get the stats standards for network packages moving soon. Are you still interested?
